requirements.txt
pandas
pytest
streamlit
python-dotenv
openai
requests
